requirements.txt
langchain
langchain_community
numpy
streamlit=1.29.0
python-dotenv
PyPDF2
transformers
sentence-transformers
llama-cpp-python
pandas 
-e . 